AppSettings + DetailPage + MapPage + RoutePage
In App Settings is the last known location added.
In The Detail page is the close button function changed. + Clean up.
In The Map Page is page changed so the last known location will be used
to load the map. And the method to create a Geopoint has been
abstracted.
In The Route Page is the calculation method added to calculate your
current GPS location and add this to the Application Data.