routes.json + SubjectResources + RoutePage + DataSource + JsonIO
routes.json has been extensive so all the known points are added. The
imageNames could not match the point.
For the subjectresources go to google.drive.nl -> 23TI2.2A3 -> Proftaak
-> SubjectResources.
The route page has now a feedback textblock. so the user knows the
application is doing something.
Datasource has got a 'corruptkey' attribute so when the routes.json has
a invalid field, you know which one it is.